# Generate script resource preview without parsing